+++ tests/unit/mocks/config.js
@@ -1,3 +1,5 @@
+var _ = require('lodash');
+
 module.exports = Config;
 
 
@@ -31,9 +33,34 @@ function Config(data) {
  *
  * @param {String} key
  *
- * @return {undefined}
+ * @return {mixed}
  */
 Config.prototype.get = function(key) {
-    return this._store[key];
+    var keys = (key || '').split(':');
+    return _.get(this._store, keys);
 };
 
+/**
+ * getOrFail
+ *
+ * @param {String} key
+ *
+ * @return {mixed}
+ */
+Config.prototype.getOrFail = function(key) {
+    var keys = (key || '').split(':');
+    var val = _.get(this._store, keys);
+    if (val === undefined) {
+        throw new Error(`Cant find config value of "${key}"`);
+    }
+    return val;
+};
+
+/**
+ * @param {Object} data
+ *
+ * @return {Config}
+ */
+Config.prototype.createLiteralProvider = function(data) {
+    return new Config(data);
+};